fix handling of editable field

+++ bld/ui/c/uivline.c
@@ -122,7 +122,7 @@ ui_event UIAPI uiveditevent( VSCREEN *vptr, VEDITLINE *editline, ui_event ui_ev 
         scroll = editline->scroll;
         if( scroll > editline->index )
             scroll = editline->index;
-        if( scroll < editline->index - editline->fldlen + 1 )
+        if( scroll < (int)( editline->index - editline->fldlen + 1 ) )
             scroll = editline->index - editline->fldlen + 1;
         editline->scroll = scroll;
         vptr->col = editline->col + editline->index - editline->scroll;
@@ -149,7 +149,7 @@ ui_event UIAPI uiveditevent( VSCREEN *vptr, VEDITLINE *editline, ui_event ui_ev 
                 scroll = editline->scroll;
                 if( scroll > editline->index )
                     scroll = editline->index;
-                if( scroll < editline->index - editline->fldlen + 1 ) {
+                if( scroll < (int)( editline->index - editline->fldlen + 1 ) ) {
                     scroll = editline->index - editline->fldlen + 1;
                 }
             } else {
